+++ src/beetl/config/__init__.py
@@ -1,0 +1,2 @@
+from .config_base import BeetlConfig, SyncConfiguration, ComparisonColumn
+from .config_v1 import BeetlConfigV1
